fixed testdata - hibernate validated schema.sql and data.sql

+++ src/main/java/pl/klugeradoslaw/mylibrary/book/Book.java
@@ -20,6 +20,6 @@ public class Book {
     private Genre genre;
     @OneToMany(mappedBy = "book")
     private List<Rating> ratings;
-    private int isbn;
+    private Long isbn;
 
 }
